moving notes to seperate python file
